Restructure shell-token classes
Renamed classes to be more concise/accurate in description as per POSIX token definitions.
Replaced CommandToken class with generic WordToken class, and added NewlineToken class. The aim is to generalize the lexer's responsibility, and leave the heavy-lifting to the parser.
Implemented print virtual functions for each class type for debugging lexer output.